change upgradeData script to the patch
