--- Setup/UpgradeData.php
@@ -1,115 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-
-namespace Magento\Braintree\Setup;
-
-use Magento\Framework\DB\DataConverter\SerializedToJson;
-use Magento\Framework\DB\FieldDataConversionException;
-use Magento\Framework\DB\FieldDataConverterFactory;
-use Magento\Framework\DB\Select\QueryModifierFactory;
-use Magento\Framework\Setup\ModuleContextInterface;
-use Magento\Framework\Setup\ModuleDataSetupInterface;
-use Magento\Framework\Setup\UpgradeDataInterface;
-
-/**
- * Class UpgradeData
- * @package Magento\Braintree\Setup
- */
-class UpgradeData implements UpgradeDataInterface
-{
-    /**
-     * @var FieldDataConverterFactory
-     */
-    private $fieldDataConverterFactory;
-
-    /**
-     * @var QueryModifierFactory
-     */
-    private $queryModifierFactory;
-
-    /**
-     * UpgradeData constructor.
-     *
-     * @param FieldDataConverterFactory $fieldDataConverterFactory
-     * @param QueryModifierFactory $queryModifierFactory
-     */
-    public function __construct(
-        FieldDataConverterFactory $fieldDataConverterFactory,
-        QueryModifierFactory $queryModifierFactory
-    ) {
-        $this->fieldDataConverterFactory = $fieldDataConverterFactory;
-        $this->queryModifierFactory = $queryModifierFactory;
-    }
-
-    /**
-     * Upgrades data for Braintree module
-     *
-     * @param ModuleDataSetupInterface $setup
-     * @param ModuleContextInterface $context
-     * @return void
-     * @throws FieldDataConversionException
-     */
-    public function upgrade(ModuleDataSetupInterface $setup, ModuleContextInterface $context)
-    {
-        if (version_compare($context->getVersion(), '2.0.1', '<')) {
-            $this->convertSerializedDataToJson($setup);
-        }
-
-        // Hotfix to remove incorrectly stored LPM config key/values
-        if (version_compare($context->getVersion(), '4.0.1', '<')) {
-            $this->removeNullLpmAllowedMethods($setup);
-        }
-    }
-
-    /**
-     * Upgrade data to version 2.0.1, converts row data in the core_config_data table that uses the path
-     * payment/braintree/countrycreditcard from serialized to JSON
-     *
-     * @param ModuleDataSetupInterface $setup
-     * @return void
-     * @throws FieldDataConversionException
-     */
-    private function convertSerializedDataToJson(ModuleDataSetupInterface $setup)
-    {
-        $fieldDataConverter = $this->fieldDataConverterFactory->create(
-            SerializedToJson::class
-        );
-
-        $queryModifier = $this->queryModifierFactory->create(
-            'in',
-            [
-                'values' => [
-                    'path' => ['payment/braintree/countrycreditcard']
-                ]
-            ]
-        );
-
-        $fieldDataConverter->convert(
-            $setup->getConnection(),
-            $setup->getTable('core_config_data'),
-            'config_id',
-            'value',
-            $queryModifier
-        );
-    }
-
-    /**
-     * Hotfix to remove incorrectly stored config value for `payment/braintree_local_paymnet/allowed_methods`
-     *
-     * @param ModuleDataSetupInterface $setup
-     */
-    private function removeNullLpmAllowedMethods(ModuleDataSetupInterface $setup)
-    {
-        $connection = $setup->getConnection();
-        $connection->delete(
-            $setup->getTable('core_config_data'),
-            [
-                '`path`=?' => 'payment/braintree_local_payment/allowed_methods',
-                '`value` IS NULL'
-            ]
-        );
-    }
-}
